temp_old_files/requirements (6).txt
streamlit>=1.34
pandas>=2.1
openpyxl>=3.1
pillow>=10.2
requests>=2.31
